dts-admin entity和pipeline'task接口和测试类编写

+++ dts-admin/src/main/java/com/utopia/data/transfer/admin/vo/req/PipelinePairAddVo.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 public class PipelinePairAddVo {
 
 
-    private Integer pipelineId;
+    private Long pipelineId;
 
     private String sourceName;
